% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CRTgeeDR-main.R
\docType{package}
\name{CRTgeeDR}
\alias{CRTgeeDR}
\alias{CRTgeeDR-package}
\title{Doubly Robust Inverse Probability Weighted Augmented GEE estimator}
\description{
The CRTgeeDR package allows you to estimates parameters in a regression model (with possibly a link function).
It allows treatment augmentation and IPW for missing data alone.
}
\details{
The only function you're likely to need from \pkg{CRTgeeDR} is
\code{\link{geeDREstimation}}. Otherwise refer to the help documentation.
}

