% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CRTgeeDR-main.R
\name{data.sim}
\alias{data.sim}
\title{The data.sim Dataset.}
\format{A data frame with 10000 rows and 8 variables}
\description{
HIV risk of infection after STI/HIV intervention in a cluster randomized trial.
}
\details{
A dataset containing the HIV risk scores and presence of risky behaviors (yes/no) and other covarites of 10000 subjects among 100 communities.
The variables are as follows:

\itemize{
  \item IDPAT subject id
  \item CLUSTER cluster id
  \item TRT treatment status, 1 is received STI/HIV intervention
  \item X1 A covariate following a N(0,1)
  \item JOB employement status
  \item MARRIED marital status
  \item AGE age
  \item HIV.KNOW Score for HIV knowlege
  \item RELIGION religiosity score
  \item OUTCOME Binary outcome - 1 if the subject is at high risk of HIV infection, 0 otherwise. NA if missing.
  \item MISSING 1 if the ouctome is missing - 0 otherwise.
}
}

