% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{fitted.CRTgeeDR}
\alias{fitted}
\alias{fitted.CRTgeeDR}
\title{Fit CRTgeeDR object.}
\usage{
\method{fitted}{CRTgeeDR}(object, ...)
}
\arguments{
\item{object}{CRTgeeDR object}

\item{...}{ignored}
}
\description{
Fit CRTgeeDR object to a dataset
}

