% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{getCI}
\alias{getCI}
\title{Get Mean, Sd and CI for estimates from CRTgeeDR object.}
\usage{
getCI(object, nameTRT = "TRT", quantile = 1.96)
}
\arguments{
\item{object}{CRTgeeDR}

\item{nameTRT,}{character including the name of the variable of interest (often the treatment)}

\item{quantile,}{value of the normal quantile for the IC. default is 1.96 for 95\%CI.}
}
\description{
Get the estimates, standard deviations and confidence intervals from an CRTgeeDR object associated with a regressor given in argument.
}

