% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{getOMPlot}
\alias{getOMPlot}
\title{Get the observed vs fitted residuals}
\usage{
getOMPlot(object, save = FALSE, name = "plotOM", typeplot = 0)
}
\arguments{
\item{object}{CRTgeeDR}

\item{save,}{logical if TRUE the plot is saved as a pdf in the current directory}

\item{name,}{name of the plot saved as pdf}

\item{typeplot,}{integer indicating which is the adequation diagnostic plot for the PS. '0', all available in plot.glm are displayed, '1' Residuals vs Fitted, '2' Normal Q-Q,
'3' Scale-Location, '4' Cook's distance, '5' Residuals vs Leverage and '6' Cook's dist vs Leverage* h[ii] / (1 - h[ii])}
}
\description{
Get the histogram and some basic statistics for the weights used in the IPW part.
}

