% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{predict.CRTgeeDR}
\alias{predict}
\alias{predict.CRTgeeDR}
\title{Predict CRTgeeDR object.}
\usage{
\method{predict}{CRTgeeDR}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{CRTgeeDR object}

\item{newdata}{dataframe, new dataset to which the CRTgeeDRneed to be used for prediction}

\item{...}{ignored}
}
\description{
Predict CRTgeeDR object to a dataset
}

