% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.CRTgeeDR}
\alias{print}
\alias{print.CRTgeeDR}
\title{Prints CRTgeeDR object.}
\usage{
\method{print}{CRTgeeDR}(x, ...)
}
\arguments{
\item{x}{CRTgeeDR x}

\item{...}{ignored}
}
\description{
Prints CRTgeeDR object
}

