% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.CRTgeeDR}
\alias{print.summary}
\alias{print.summary.CRTgeeDR}
\title{Print the summarizing CRTgeeDR object.}
\usage{
\method{print}{summary.CRTgeeDR}(x, ...)
}
\arguments{
\item{x}{summary.CRTgeeDR x}

\item{...}{ignored}
}
\description{
Print Summary CRTgeeDR object
}

