% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{summary.CRTgeeDR}
\alias{summary}
\alias{summary.CRTgeeDR}
\title{Summarizing CRTgeeDR object.}
\usage{
\method{summary}{CRTgeeDR}(object, ...)
}
\arguments{
\item{object}{CRTgeeDR object}

\item{...}{ignored}
}
\description{
Summary CRTgeeDR object
}

